Rename experiment description editor file
Move the experiment description editor to its metadata-oriented filename before broadening the modal so later changes review as edits instead of a delete and re-add.
